refactor: move message creation endpoint
move message creation from DomainController to messageController.

related to DOP-1171
