+++ Doppler.PushContact/Controllers/MessageController.cs
@@ -49,5 +49,29 @@ namespace Doppler.PushContact.Controllers
                 MessageId = messageId
             });
         }
+
+        [HttpPost]
+        [Route("message")]
+        public async Task<IActionResult> CreateMessage([FromBody] MessageBody message)
+        {
+            try
+            {
+                // TODO: analyze remotion of validation for the title and body, it's being doing during model binding with annotations.
+                _messageSender.ValidateMessage(message.Title, message.Body, message.OnClickLink, message.ImageUrl);
+            }
+            catch (ArgumentException argExc)
+            {
+                return UnprocessableEntity(argExc.Message);
+            }
+
+            var messageId = Guid.NewGuid();
+
+            await _messageRepository.AddAsync(messageId, message.Domain, message.Title, message.Body, message.OnClickLink, 0, 0, 0, message.ImageUrl);
+
+            return Ok(new MessageResult
+            {
+                MessageId = messageId
+            });
+        }
     }
 }
